requirements.txt
Flask==2.3.2
Flask-RESTX==0.5.1
python-dotenv==1.0.0
pytest==7.3.1
